+++ sdk/core/azure-core/src/main/java/com/azure/core/credential/KeyCredential.java
@@ -3,10 +3,18 @@
 
 package com.azure.core.credential;
 
+import com.azure.core.util.logging.ClientLogger;
+
+import java.util.Objects;
+
 /**
- * Represents a credential that uses a key to authenticate to an Azure Service.
+ * Represents a credential that uses a key to authenticate.
  */
-public final class AzureKeyCredential extends KeyCredential {
+public class KeyCredential {
+    // AzureKeyCredential is a commonly used credential type, use a static logger.
+    private static final ClientLogger LOGGER = new ClientLogger(KeyCredential.class);
+    private volatile String key;
+
     /**
      * Creates a credential that authorizes request with the given key.
      *
@@ -14,8 +22,13 @@ public final class AzureKeyCredential extends KeyCredential {
      * @throws NullPointerException If {@code key} is {@code null}.
      * @throws IllegalArgumentException If {@code key} is an empty string.
      */
-    public AzureKeyCredential(String key) {
-        super(key);
+    public KeyCredential(String key) {
+        Objects.requireNonNull(key, "'key' cannot be null.");
+        if (key.isEmpty()) {
+            throw LOGGER.logExceptionAsError(new IllegalArgumentException("'key' cannot be empty."));
+        }
+
+        this.key = key;
     }
 
     /**
@@ -23,9 +36,8 @@ public final class AzureKeyCredential extends KeyCredential {
      *
      * @return The key being used to authorize requests.
      */
-    @Override
     public String getKey() {
-        return super.getKey();
+        return key;
     }
 
     /**
@@ -36,9 +48,13 @@ public final class AzureKeyCredential extends KeyCredential {
      * @throws NullPointerException If {@code key} is {@code null}.
      * @throws IllegalArgumentException If {@code key} is an empty string.
      */
-    @Override
-    public AzureKeyCredential update(String key) {
-        super.update(key);
+    public KeyCredential update(String key) {
+        Objects.requireNonNull(key, "'key' cannot be null.");
+        if (key.isEmpty()) {
+            throw LOGGER.logExceptionAsError(new IllegalArgumentException("'key' cannot be empty."));
+        }
+
+        this.key = key;
         return this;
     }
 }