+++ sdk/core/azure-core/src/main/java/com/azure/core/http/policy/AzureKeyCredentialPolicy.java
@@ -4,14 +4,10 @@
 package com.azure.core.http.policy;
 
 import com.azure.core.credential.AzureKeyCredential;
-import com.azure.core.http.HttpHeaderName;
-import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.HttpPipelineCallContext;
 import com.azure.core.http.HttpPipelineNextPolicy;
 import com.azure.core.http.HttpPipelineNextSyncPolicy;
 import com.azure.core.http.HttpResponse;
-import com.azure.core.util.FluxUtil;
-import com.azure.core.util.logging.ClientLogger;
 import reactor.core.publisher.Mono;
 
 import java.util.Objects;
@@ -22,13 +18,7 @@ import java.util.Objects;
  * Requests sent with this pipeline policy are required to use {@code HTTPS}. If the request isn't using {@code HTTPS}
  * an exception will be thrown to prevent leaking the key.
  */
-public final class AzureKeyCredentialPolicy implements HttpPipelinePolicy {
-    // AzureKeyCredentialPolicy can be a commonly used policy, use a static logger.
-    private static final ClientLogger LOGGER = new ClientLogger(AzureKeyCredentialPolicy.class);
-    private final HttpHeaderName name;
-    private final AzureKeyCredential credential;
-    private final String prefix;
-
+public final class AzureKeyCredentialPolicy extends KeyCredentialPolicy {
     /**
      * Creates a policy that uses the passed {@link AzureKeyCredential} to set the specified header name.
      *
@@ -38,7 +28,7 @@ public final class AzureKeyCredentialPolicy implements HttpPipelinePolicy {
      * @throws IllegalArgumentException If {@code name} is empty.
      */
     public AzureKeyCredentialPolicy(String name, AzureKeyCredential credential) {
-        this(name, credential, null);
+        super(name, credential, null);
     }
 
     /**
@@ -54,48 +44,16 @@ public final class AzureKeyCredentialPolicy implements HttpPipelinePolicy {
      * @throws IllegalArgumentException If {@code name} is empty.
      */
     public AzureKeyCredentialPolicy(String name, AzureKeyCredential credential, String prefix) {
-        this(validateName(name), Objects.requireNonNull(credential, "'credential' cannot be null."), prefix);
-    }
-
-    private static HttpHeaderName validateName(String name) {
-        Objects.requireNonNull(name, "'name' cannot be null.");
-        if (name.isEmpty()) {
-            throw LOGGER.logExceptionAsError(new IllegalArgumentException("'name' cannot be empty."));
-        }
-
-        return HttpHeaderName.fromString(name);
-    }
-
-    AzureKeyCredentialPolicy(HttpHeaderName name, AzureKeyCredential credential, String prefix) {
-        this.name = name;
-        this.credential = credential;
-        this.prefix = prefix != null ? prefix.trim() : null;
+        super(name, Objects.requireNonNull(credential, "'credential' cannot be null."), prefix);
     }
 
     @Override
     public Mono<HttpResponse> process(HttpPipelineCallContext context, HttpPipelineNextPolicy next) {
-        if ("http".equals(context.getHttpRequest().getUrl().getProtocol())) {
-            return FluxUtil.monoError(LOGGER,
-                new IllegalStateException("Key credentials require HTTPS to prevent leaking the key."));
-        }
-
-        setCredential(context.getHttpRequest().getHeaders());
-        return next.process();
+        return super.process(context, next);
     }
 
     @Override
     public HttpResponse processSync(HttpPipelineCallContext context, HttpPipelineNextSyncPolicy next) {
-        if ("http".equals(context.getHttpRequest().getUrl().getProtocol())) {
-            throw LOGGER.logExceptionAsError(
-                new IllegalStateException("Key credentials require HTTPS to prevent leaking the key."));
-        }
-
-        setCredential(context.getHttpRequest().getHeaders());
-        return next.processSync();
-    }
-
-    void setCredential(HttpHeaders headers) {
-        String credential = this.credential.getKey();
-        headers.set(name, (prefix == null) ? credential : prefix + " " + credential);
+        return super.processSync(context, next);
     }
 }