+++ sdk/core/azure-core/src/main/java/com/azure/core/http/policy/KeyCredentialPolicy.java
@@ -1,16 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.ai.openai.implementation;
+package com.azure.core.http.policy;
 
-import com.azure.ai.openai.models.NonAzureOpenAIKeyCredential;
+import com.azure.core.credential.KeyCredential;
 import com.azure.core.http.HttpHeaderName;
 import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.HttpPipelineCallContext;
 import com.azure.core.http.HttpPipelineNextPolicy;
 import com.azure.core.http.HttpPipelineNextSyncPolicy;
 import com.azure.core.http.HttpResponse;
-import com.azure.core.http.policy.HttpPipelinePolicy;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
 import reactor.core.publisher.Mono;
@@ -18,35 +17,43 @@ import reactor.core.publisher.Mono;
 import java.util.Objects;
 
 /**
- * Pipeline policy that uses an {@link NonAzureOpenAIKeyCredential} to set the authorization key for a request to
- * an HTTP request.
+ * Pipeline policy that uses an {@link KeyCredential} to set the authorization key for a request.
  * <p>
  * Requests sent with this pipeline policy are required to use {@code HTTPS}. If the request isn't using {@code HTTPS}
  * an exception will be thrown to prevent leaking the key.
  */
-public class NonAzureOpenAIKeyCredentialPolicy implements HttpPipelinePolicy {
-    private static final ClientLogger LOGGER = new ClientLogger(NonAzureOpenAIKeyCredentialPolicy.class);
+public class KeyCredentialPolicy implements HttpPipelinePolicy {
+    private static final ClientLogger LOGGER = new ClientLogger(KeyCredentialPolicy.class);
     private final HttpHeaderName name;
-    private final NonAzureOpenAIKeyCredential credential;
-
+    private final KeyCredential credential;
     private final String prefix;
 
     /**
-     * Creates a policy that uses the passed {@link NonAzureOpenAIKeyCredential} to set the specified header name.
-     * <p>
-     * The {@code prefix} will be applied before the {@link NonAzureOpenAIKeyCredential#getKey()} when setting the
-     * header. A space will be inserted between {@code prefix} and credential.
+     * Creates a policy that uses the passed {@link KeyCredential} to set the specified header name.
      *
-     * @param name The name of the key header that will be set to {@link NonAzureOpenAIKeyCredential#getKey()}.
-     * @param credential The {@link NonAzureOpenAIKeyCredential} containing the authorization key to use.
-     * @param prefix The prefix to apply before the credential, for example "Bearer" or "Basic".
+     * @param name The name of the key header that will be set to {@link KeyCredential#getKey()}.
+     * @param credential The {@link KeyCredential} containing the authorization key to use.
      * @throws NullPointerException If {@code name} or {@code credential} is {@code null}.
      * @throws IllegalArgumentException If {@code name} is empty.
      */
-    public NonAzureOpenAIKeyCredentialPolicy(String name, NonAzureOpenAIKeyCredential credential, String prefix) {
-        this.name = validateName(name);
-        this.credential = Objects.requireNonNull(credential, "'credential' cannot be null.");
-        this.prefix = prefix != null ? prefix.trim() : null;
+    public KeyCredentialPolicy(String name, KeyCredential credential) {
+        this(name, credential, null);
+    }
+
+    /**
+     * Creates a policy that uses the passed {@link KeyCredential} to set the specified header name.
+     * <p>
+     * The {@code prefix} will be applied before the {@link KeyCredential#getKey()} when setting the header. A
+     * space will be inserted between {@code prefix} and credential.
+     *
+     * @param name The name of the key header that will be set to {@link KeyCredential#getKey()}.
+     * @param credential The {@link KeyCredential} containing the authorization key to use.
+     * @param prefix The prefix to apply before the credential, for example "SharedAccessKey credential".
+     * @throws NullPointerException If {@code name} or {@code credential} is {@code null}.
+     * @throws IllegalArgumentException If {@code name} is empty.
+     */
+    public KeyCredentialPolicy(String name, KeyCredential credential, String prefix) {
+        this(validateName(name), Objects.requireNonNull(credential, "'credential' cannot be null."), prefix);
     }
 
     private static HttpHeaderName validateName(String name) {
@@ -56,6 +63,12 @@ public class NonAzureOpenAIKeyCredentialPolicy implements HttpPipelinePolicy {
         }
 
         return HttpHeaderName.fromString(name);
+    }
+
+    KeyCredentialPolicy(HttpHeaderName name, KeyCredential credential, String prefix) {
+        this.name = name;
+        this.credential = credential;
+        this.prefix = prefix != null ? prefix.trim() : null;
     }
 
     @Override