+++ sdk/core/azure-core/src/test/java/com/azure/core/http/policy/AzureKeyCredentialPolicyTests.java
@@ -26,10 +26,10 @@ public class AzureKeyCredentialPolicyTests {
     private static Stream<Arguments> setCredentialSupplier() {
         AzureKeyCredential credential = new AzureKeyCredential("asecret");
         return Stream.of(
-            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION, credential, null), "asecret"),
-            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION, credential, "SharedKeyCredential"),
+            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION.toString(), credential, null), "asecret"),
+            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION.toString(), credential, "SharedKeyCredential"),
                 "SharedKeyCredential asecret"),
-            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION, credential, "SharedKeyCredential "),
+            Arguments.of(new AzureKeyCredentialPolicy(HttpHeaderName.AUTHORIZATION.toString(), credential, "SharedKeyCredential "),
                 "SharedKeyCredential asecret")
         );
     }