+++ sdk/openai/azure-ai-openai/src/samples/java/com/azure/ai/openai/impl/ReadmeSamples.java
@@ -18,8 +18,8 @@ import com.azure.ai.openai.models.CompletionsUsage;
 import com.azure.ai.openai.models.EmbeddingItem;
 import com.azure.ai.openai.models.Embeddings;
 import com.azure.ai.openai.models.EmbeddingsOptions;
-import com.azure.ai.openai.models.NonAzureOpenAIKeyCredential;
 import com.azure.core.credential.AzureKeyCredential;
+import com.azure.core.credential.KeyCredential;
 import com.azure.core.credential.TokenCredential;
 import com.azure.core.http.ProxyOptions;
 import com.azure.core.util.HttpClientOptions;
@@ -61,7 +61,7 @@ public final class ReadmeSamples {
     public void createNonAzureSyncClientWithApiKey() {
         // BEGIN: readme-sample-createNonAzureOpenAISyncClientApiKey
         OpenAIClient client = new OpenAIClientBuilder()
-            .credential(new NonAzureOpenAIKeyCredential("{openai-secret-key}"))
+            .credential(new KeyCredential("{openai-secret-key}"))
             .buildClient();
         // END: readme-sample-createNonAzureOpenAISyncClientApiKey
     }
@@ -69,7 +69,7 @@ public final class ReadmeSamples {
     public void createNonAzureAsyncClientWithApiKey() {
         // BEGIN: readme-sample-createNonAzureOpenAIAsyncClientApiKey
         OpenAIAsyncClient client = new OpenAIClientBuilder()
-            .credential(new NonAzureOpenAIKeyCredential("{openai-secret-key}"))
+            .credential(new KeyCredential("{openai-secret-key}"))
             .buildAsyncClient();
         // END: readme-sample-createNonAzureOpenAIAsyncClientApiKey
     }