+++ sdk/openai/azure-ai-openai/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 ### Breaking Changes
+- Removed `NonAzureOpenAIKeyCredential` and `NonAzureOpenAIKeyCredentialPolicy`. Replace by Azure Core `KeyCredential` and `KeyCredentialPolicy`.
 
 ### Bugs Fixed